requirements.txt
biopython
numpy
pandas
scikit-learn
sentencepiece
torch
tqdm
transformers
